requirements.txt
flask
streamlit
pypdf
cryptography
google-generativeai
requests
python-dotenv